requirements.txt
httplib2
django
django-cors-headers
Pillow
umemcache
git+git://github.com/tilejet/django-memcached-pool.git@tilejet
gunicorn
greenlet
gevent
git+git://github.com/tilejet/celery.git@umemcache
geojson
pymongo
iso8601
django-jquery
pinax-theme-bootstrap
django_forms_bootstrap
git+git://github.com/tilejet/tilejet-util@master
git+git://github.com/tilejet/tilejet-logs@master
git+git://github.com/tilejet/tilejet-stats@master
git+git://github.com/tilejet/tilejet-cache@master
